# Conflicts:
#	apps/form_builder/app/routes/auth/github.tsx
#	apps/form_builder/app/routes/auth/github_.callback.tsx
#	apps/form_builder/package.json
#	pnpm-lock.yaml

+++ packages/registry/src/components/parish.ts
@@ -19,7 +19,7 @@ export const Parish: SelectPrimitive = {
   ],
   multiple: false,
   ui: {
-    width: "short",
+    width: "long",
   },
   validations: {
     required: {
